persistence

+++ daemon/lf/LogicFaceSystem.go
@@ -106,7 +106,7 @@ func (l *LogicFaceSystem) doFaceClean() {
 		if v.state == false {
 			common2.LogInfo("remove logicface id = ", v.LogicFaceId)
 			l.destroyFace(k, v)
-		} else if v.expireTime < curTime { // logicFace已经超时
+		} else if v.expireTime < curTime && v.Persistence == 0{ // logicFace已经超时
 			v.Shutdown()        // 调用shutdown关闭logicFace
 			l.destroyFace(k, v) // 将logicFace从全局logicFaceTable中删除
 		}
